+++ src/Resources/contao/classes/ls_shop_orderMessages.php
@@ -380,6 +380,23 @@ class ls_shop_orderMessages
 			if ($arrMessageToSendAndSave['receiverBccAddress']) {
 				$objEmail->sendBcc($arrMessageToSendAndSave['receiverBccAddress']);
 			}
+
+            if(!empty($arrMessageModel['mailerTransport'])) {
+                // Set the transport
+                $objEmail->addHeader('X-Transport', $arrMessageModel['mailerTransport']);
+            }
+
+            // Get the "reply to" address
+            if (!empty($arrMessageModel['replyTo'])) {
+                $replyTo = $arrMessageModel['replyTo'];
+
+                // Add the name
+                if (!empty($arrMessageModel['replyToName'])) {
+                    $replyTo = '"' . $arrMessageModel['replyToName'] . '" <' . $replyTo . '>';
+                }
+
+                $objEmail->replyTo($replyTo);
+            }
 			
 			try {
 				$objEmail->sendTo($arrMessageToSendAndSave['receiverMainAddress']);