+++ CHANGELOG.md
@@ -3,3 +3,4 @@ Leading Systems Contao Merconis bundle changelog
 
 ### 5.1.0 (2025-02-04)
  * add feature productCode and alias to inserttag {{shopProductProperty::*}}, {{shopProductOutput::*}}
+ * add Mailer Transport for Shopmessages